integrated nba export

+++ include/LTL/Structures/ProductState.h
@@ -50,7 +50,7 @@ namespace LTL::Structures {
         }
 
         [[nodiscard]] bool markingEqual(const PetriEngine::MarkVal *rhs) const {
-            for (int i = 0; i < buchi_state_idx; ++i) {
+            for (size_t i = 0; i < buchi_state_idx; ++i) {
                 if (marking()[i] != rhs[i]) {
                     return false;
                 }
@@ -59,7 +59,7 @@ namespace LTL::Structures {
         }
 
         bool operator==(const ProductState &rhs) const {
-            for (int i = 0; i <= buchi_state_idx; ++i) {
+            for (size_t i = 0; i <= buchi_state_idx; ++i) {
                 if (marking()[i] != rhs.marking()[i]) {
                     return false;
                 }
